Rename accelelerators
